introduced base diagram object
