+++ src/compas_ags/rhino/__init__.py
@@ -23,32 +23,26 @@ Objects
     :toctree: generated/
     :nosignatures:
 
+    DiagramObject
     FormObject
-
-Scenes
-======
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
 
 """
 from __future__ import absolute_import
 
-from compas_ags.diagrams import FormDiagram
+from compas_ags.diagrams import FormDiagram  # noqa: F401
 from compas_ags.diagrams import ForceDiagram  # noqa: F401
 
 from .diagramartist import DiagramArtist  # noqa: F401
-from .formartist import FormArtist
+from .formartist import FormArtist  # noqa: F401
 from .forceartist import ForceArtist  # noqa: F401
 
-from .formobject import FormObject
+from .diagramobject import DiagramObject  # noqa: F401
+from .formobject import FormObject  # noqa: F401
 
-from .scene import Scene  # noqa: F401
-from .scene import SceneObject
+# from .scene import Scene  # noqa: F401
+# from .scene import SceneObject
 
-
-SceneObject.register(FormDiagram, FormObject, FormArtist)
+# SceneObject.register(FormDiagram, FormObject, FormArtist)
 
 
 __all__ = [name for name in dir() if not name.startswith('_')]
